refactor: separate component directory
