+++ src/features/core/components/atomic/avatar.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react';
 
-import { cn } from '../utils/cn';
+import { cn } from '../../utils/cn';
 
 interface AvatarProps extends React.HTMLAttributes<HTMLDivElement> {
   seed?: number;
