+++ src/features/core/components/atomic/icon-button.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../utils/cn';
+import { cn } from '../../utils/cn';
 import { Button, ButtonProps } from './button';
 
 export type IconButtonProps = Omit<ButtonProps, 'prefixIcon' | 'suffixIcon'> & {
