+++ src/features/core/components/atomic/loading-overlay.tsx
@@ -1,6 +1,6 @@
 import { PropsWithChildren } from 'react';
 
-import { cn } from '../utils/cn';
+import { cn } from '../../utils/cn';
 
 export function LoadingOverlay({
   children,
