+++ src/features/core/components/log/log-click.tsx
