+++ src/features/core/index.ts
@@ -1,26 +1,31 @@
-export * from './components/back-button';
-export * from './components/backdrop';
-export * from './components/bottom-sheet';
-export * from './components/button';
-export * from './components/checkbox';
-export * from './components/funnel-layout';
-export * from './components/input';
-export * from './components/label';
-export * from './components/theme-switcher';
-export * from './components/loading-ellipse';
-export * from './components/loading-overlay';
-export * from './components/dialog';
-export * from './components/multi-state-switch';
-export * from './components/password-input';
-export * from './components/switch';
-export * from './components/avatar';
-export * from './components/file-upload';
-export * from './components/modal';
-export * from './components/icon-button';
-export * from './components/copy-input';
-export * from './components/toast';
-export * from './components/swipe-card';
-export * from './components/log-click';
+// Atomic components
+export * from './components/atomic/avatar';
+export * from './components/atomic/back-button';
+export * from './components/atomic/backdrop';
+export * from './components/atomic/button';
+export * from './components/atomic/checkbox';
+export * from './components/atomic/icon-button';
+export * from './components/atomic/input';
+export * from './components/atomic/label';
+export * from './components/atomic/loading-ellipse';
+export * from './components/atomic/loading-overlay';
+export * from './components/atomic/password-input';
+export * from './components/atomic/switch';
+export * from './components/atomic/theme-switcher';
+
+// Compound components
+export * from './components/compound/bottom-sheet';
+export * from './components/compound/copy-input';
+export * from './components/compound/dialog';
+export * from './components/compound/file-upload';
+export * from './components/compound/funnel-layout';
+export * from './components/compound/modal';
+export * from './components/compound/multi-state-switch';
+export * from './components/compound/swipe-card';
+export * from './components/compound/toast';
+
+// Log components
+export * from './components/log/log-click';
 
 export * from './services/api';
 export * from './services/amplitude';
